app/requirements.txt
flask==3.0.0
line-bot-sdk>=3.17.1
python-dotenv==1.0.0
boto3>=1.37.24
requests>=2.32.3
pillow>=9.2.0,<11.0
pytest==8.0.0
pytest-mock==3.12.0
tenacity==8.2.3
python-dateutil==2.8.2
moto==4.2.14
pytest-cov==4.1.0
psutil>=5.4.8,<6.0.0
nest-asyncio>=1.5.8
langchain>=0.1.0
langgraph>=0.0.20
langgraph-checkpoint-aws>=0.0.1
langchain-aws>=0.0.1
pandas>=2.0.0